refactor batch and rt processing, fix requires next step did not respect realtime flag on message
