+++ app-server/src/signals/mod.rs
@@ -12,6 +12,7 @@ pub mod prompts;
 pub mod provider;
 pub mod queue;
 pub mod realtime_api;
+pub mod response_processor;
 pub mod spans;
 pub mod submissions_consumer;
 pub mod tools;
